Trunk reintegration (9265 - 9321)
git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/branches/3.9.test1@9322 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de

+++ src/55_timetables/TimetableGenerator.cpp
@@ -642,7 +642,7 @@ namespace synthese
 								seqEnd - itColStart == lastColSet.second - lastColSet.first
 							){
 								bool ok(true);
-								for(size_t curSetColRank(0); curSetColRank<=seqEnd - itColStart; ++curSetColRank)
+								for(size_t curSetColRank(0); curSetColRank<=(size_t)(seqEnd - itColStart); ++curSetColRank)
 								{
 									if(	!(itColStart+curSetColRank)->isLike(*(lastColSet.first + curSetColRank), delta)
 									){
